requirements.txt
PyPDF2>=3.0.0
gspread>=5.7.0
oauth2client>=4.1.3
pandas>=1.5.0
python-dotenv>=0.20.0
langchain-openai>=0.0.1
streamlit>=1.20.0
tenacity>=9.0.0,<10.0.0
protobuf>=5.26.1,<6.0
langchain-google-genai>=2.1.3
langchain-core>=0.3.52,<0.4.0
mysql-connector-python
bs4
xlsxwriter
openpyxl